+++ src/VBox/Main/xml/Settings.cpp
@@ -2478,7 +2478,7 @@ void CaptureSettings::applyDefaults(void)
     mapScreens.clear();
 
     CaptureScreenSettings screenSettings; /* Apply default settings. */
-    for (unsigned long l = 0; l < SchemaDefs::MaxGuestMonitors; l++)
+    for (uint32_t l = 0; l < SchemaDefs::MaxGuestMonitors; l++)
     {
         mapScreens[l] = screenSettings;
     }
@@ -2516,7 +2516,7 @@ bool CaptureSettings::operator==(const CaptureSettings &d) const
         return false;
 
     CaptureScreenMap::const_iterator itScreen = mapScreens.begin();
-    unsigned long i = 0;
+    uint32_t i = 0;
     while (itScreen != mapScreens.end())
     {
         CaptureScreenMap::const_iterator itScreenThat = d.mapScreens.find(i);
